Little cosmetic changes needed
First make sure you delete the instructions. I did that for you mostly, but the reflection part. Don't forget to do that! 
Under bulletpoint 3, the whole thing is very confusing, if you could separate the actual code from the information that would help a lot! Maybe put in the "grey triangle" only the actual codes.